+++ test/auth/oauth.spec.ts
@@ -184,7 +184,7 @@ describe("OpenSeaOAuth", () => {
     const roleScopes = scopes.filter(scope =>
       scope.startsWith("urn:zitadel:iam:org:project:role:"),
     )
-    expect(roleScopes).toHaveLength(10)
+    expect(roleScopes).toHaveLength(12)
     expect(roleScopes).toContain(
       "urn:zitadel:iam:org:project:role:read:favorites",
     )